refactor(jinko): align package layout with upstream convention
Move internal/cmd → internal/cli to match the layout the upstream
generator produces (verified by .github/scripts/verify-skill/verify_skill.py
which expects internal/cli/*.go).

Also registers the focused /pp-jinko skill mirror under cli-skills/
via tools/generate-skills.
