+++ library/travel/jinko/internal/cli/find_destination.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"context"
