+++ library/travel/jinko/internal/cli/find_flight.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"context"
