+++ library/travel/jinko/internal/cli/flight_search.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"context"
