+++ library/travel/jinko/internal/cli/hotel_details.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"context"
