+++ library/travel/jinko/internal/cli/trip_status.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"context"
